login func made now have to sync it up with frontend

+++ controller/blogs.js
@@ -1,11 +1,19 @@
 const blogRouter = require('express').Router();
 const Blog = require('../models/note');
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
+
+
+
+
+//To fetch all the data
 blogRouter.get('/', async (request, response) => {
   const blogs= await Blog.find({})
   response.json(blogs)
   })
 
+//TO fetch json data of one entry
 blogRouter.get('/:id', async(req,res)=>{
   const iD= req.params.id;
 
@@ -20,23 +28,44 @@ blogRouter.get('/:id', async(req,res)=>{
   
 })
 
-  blogRouter.post("/", async (request, response) => {
-    const body = request.body;
-    const user = await User.find(body.userId);
+//Post a blog
+  blogRouter.post("/", async (req, response,next) => {
+    const body = req.body;
+    // const decoded= jwt.verify(request.token,process.env.secret_jwt);
+    // if(!decoded.id){
+    //   return response.status(401).json({error:'token invalid'})
+
+    // }
+
+    // const user = await User.findById(decoded.id);
+    
     const blog = new Blog({
-      user:user.id,
+      user:req.user.id,
       ...body
     });
     const savedNote = await blog.save();
+    req.user.blogs= req.user.blogs.concat(blog.id)
+    await req.user.save()
     response.status(201).json(savedNote);
   });
 
   blogRouter.delete('/:id',async(req,res)=>{
     const iD= req.params.id;
-    console.log(iD)
+    const decode = jwt.verify(req.token,process.env.secret_jwt)
+    if(!(decode.id)) {
+      res.status(402).json({error:'invalid token'})
+      
+    }
+    const blog= await Blog.findById(iD)
+    console.log(decode.id,blog.user)
+    if(decode.id === blog.user.toString() ){
+      await Blog.findByIdAndDelete(iD);
+      res.status(204).end()
+    }else{
+      res.status(400).json({error:'you messed up bro'})
+    }
     
-    await Blog.findByIdAndDelete(iD);
-    res.status(204).end()
+    
     
   })
   blogRouter.put('/:id',async(req,res)=>{
